Added an index.ts for all class exports
