+++ core/graphics/index.ts
@@ -2,4 +2,13 @@ export { Graphics, batchProcessGraphics, initializeGraphics, updateGraphics } fr
 
 export { Lcd } from './lcd';
 
-export { Palette, initializePalette, writeColorPaletteToMemory } from './palette';
+export {
+  Palette,
+  initializePalette,
+  writeColorPaletteToMemory,
+  getMonochromeColorFromPalette,
+  getRgbColorFromPalette,
+  getColorComponentFromRgb
+} from './palette';
+
+export { getTileDataAddress, drawPixelsFromLineOfTile } from './tiles';
