+++ core/sound/channel1.ts
@@ -5,7 +5,7 @@
 // http://gbdev.gg8.se/wiki/articles/Gameboy_sound_hardware#Frequency_Sweep
 
 import { isDutyCycleClockPositiveOrNegativeForWaveform } from './duty';
-import { Cpu } from '../cpu/cpu';
+import { Cpu } from '../cpu/index';
 import {
   eightBitLoadFromGBMemory,
   eightBitStoreIntoGBMemory,
