+++ core/sound/channel4.ts
@@ -4,7 +4,7 @@
 // http://gbdev.gg8.se/wiki/articles/Gameboy_sound_hardware#Noise_Channel
 
 import { isDutyCycleClockPositiveOrNegativeForWaveform } from './duty';
-import { Cpu } from '../cpu/cpu';
+import { Cpu } from '../cpu/index';
 import {
   eightBitLoadFromGBMemory,
   eightBitStoreIntoGBMemoryWithTraps,
